+++ CHANGES.md
@@ -1,3 +1,7 @@
+# 2015-06-15
+
+* Added update-all-external-links.rb which opens all the worksheets in a folder in turn, updates their external links, then cloese them.
+
 # 2015-05-07
 
 * Add propose-replacements-for-external-links-for-versioned-files.rb to help alter references when eliminating version numbers on individual worksheets